feat(blog): make blogcard2 common blogcard component
