+++ src/components/blog/BlogCard2.vue
@@ -1,0 +1,114 @@
+<script setup lang="ts">
+import type { Blog } from "@/types/blog";
+import { generateAuthorLink } from "@/utils/redirection";
+import CustomBlurImage from "../core/CustomBlurImage.vue";
+import { getInitials } from "@/utils/getInitials";
+
+const props = defineProps<{
+  blog: Blog;
+  type: string;
+}>();
+
+const getImageUrl = ({ image }: Blog) =>
+  image?.formats?.medium?.url ??
+  image?.formats?.large?.url ??
+  image?.formats?.small?.url ??
+  image?.url ??
+  "";
+</script>
+
+<template>
+    <a
+      :key="blog.slug"
+      :href="`/${type}/${blog.slug}`"
+      class="group"
+    >
+      <div
+        class="rounded-lg overflow-hidden border border-border hover:border-primary-purple transition-colors flex flex-col h-full"
+      >
+        <div class="relative h-48 overflow-hidden">
+          <CustomBlurImage :image="getImageUrl(blog)" :altText="blog.title" />
+          <!-- <img
+              :src="getImageUrl(blog)"
+              :alt="blog.title"
+              class="object-cover transition-transform group-hover:scale-105 w-full h-full absolute inset-0"
+              />  -->
+          <div className="absolute top-3 left-3">
+            <div
+              className="bg-primary-purple text-white border-none rounded-full p-1 text-xs"
+            >
+              Openobserve
+            </div>
+          </div>
+        </div>
+        <div class="p-2 flex flex-col flex-grow">
+          <div class="flex flex-col justify-start md:items-start mb-2">
+            <div class="flex flex-wrap gap-2 mb-2">
+              <a
+                v-for="tag in blog?.categories?.slice(0, 2)"
+                :key="tag.slug"
+                :href="`/blog/tag/${tag.slug}`"
+                class="bg-light-gray text-primary-gray text-sm font-medium rounded-full px-3 py-1 capitalize"
+              >
+                {{ tag.name }}
+              </a>
+            </div>
+            <h2 class="text-lg font-semibold mb-2">
+              {{ blog?.title }}
+            </h2>
+            <p class="text-primary-gray mb-4 flex-grow text-base line-clamp-3">
+              {{ blog?.description }}
+            </p>
+            <div class="mt-auto w-full">
+              <div class="flex items-center flex-wrap gap-4">
+                <!-- Avatars -->
+                <div class="flex -space-x-3">
+                  <div
+                    v-for="it in blog.authors"
+                    :key="it.name"
+                    class="relative rounded-full border-2 border-white/80"
+                  >
+                    <div
+                      class="h-10 w-10 rounded-full overflow-hidden bg-purple-700 flex items-center justify-center text-white text-sm font-semibold"
+                    >
+                      <img
+                        v-if="it?.image?.url"
+                        :src="it.image.url"
+                        :alt="it.name"
+                        class="h-full w-full object-cover"
+                      />
+                      <span v-else>{{ getInitials(it.name) }}</span>
+                    </div>
+                  </div>
+                </div>
+
+                <!-- Author Names -->
+                <div class="flex flex-col gap-1">
+                  <address
+                    class="not-italic text-black flex flex-wrap items-center gap-1"
+                  >
+                    <template
+                      v-for="(it, index) in blog.authors"
+                      :key="`name-${it.name}`"
+                    >
+                      <span class="font-medium">{{ it.name }}</span>
+                      <span
+                        v-if="index < blog.authors.length - 1"
+                        class="text-black text-xs"
+                        >,</span
+                      >
+                    </template>
+                  </address>
+                  <span
+                    v-if="blog?.publishDate"
+                    class="text-sm text-primary-gray"
+                    >{{ blog?.publishDate }}</span
+                  >
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </a>
+</template>
